audiomanager: fix orphan focus holder when using ext focus policy
When an focus management is delegate to an external Audio Policy, the
answer to the request is expected within 250ms. If the timer expires,
it is rejected. Timer is armed on client side.
It may lead to orphan focus owner as ext AudioPolicy may be temporary
stuck and would manage the request anyway.

This CL fixes the orphan owner by abandonning any timed out request.

Bug: 400752237
Test: m

Google: 3527210
Change-Id: I0b4e60475e83b8faa58068fe2b7753c6fbd6df7c
Signed-off-by: François Gaffie <francois.gaffie@ampere.cars>
Signed-off-by: ryzenforce990 <ryzenforce990@users.noreply.github.com>